featurize ww and codemod apply
