+++ apps/frontend/app/(website)/studio/main/PageBottomPane/Components/index.ts
@@ -1,4 +1,4 @@
-export * from "./CodeSnippets";
+export * from "./Snippets/CodeSnippets";
 export * from "./WarningTexts";
 export * from "./SnippedHeader";
 export * from "./side-components";
