+++ apps/frontend/app/(website)/studio/src/hooks/useAPI.ts
@@ -1,9 +1,12 @@
 import apiClient from "@/utils/apis/client";
 import { useAuth } from "@clerk/nextjs";
 import { mockedEndpoints } from "@shared/mocks";
+import { isServer } from "@studio/config";
 import type { AxiosResponse } from "axios";
 const shouldUseMocks =
-  process.env.NODE_ENV === "development" && localStorage.getItem("useMocks");
+  !isServer &&
+  process.env.NODE_ENV === "development" &&
+  localStorage.getItem("useMocks");
 const mockified = (
   verb: "put" | "get" | "post",
   endpoint: string | ((x: any) => string),
