+++ apps/frontend/package.json
@@ -7,6 +7,7 @@
     "dev": "next",
     "format": "npx prettier --write . --ignore-path .gitignore",
     "biome": "npx @biomejs/biome lint --error-on-warnings ./",
+    "biome:fix": "npx @biomejs/biome lint --apply ./",
     "biome:errors": "npx @biomejs/biome lint --diagnostic-level=error ./",
     "lint": "next lint -- --ignore-path .gitignore",
     "lint:fix": "npm run format && npm run lint -- --fix",
